feat: Complete US-1

+++ db/migrate/20240301043534_create_bulk_discounts.rb
@@ -3,7 +3,7 @@ class CreateBulkDiscounts < ActiveRecord::Migration[7.1]
     create_table :bulk_discounts do |t|
       t.references :merchant, null: false, foreign_key: true
       t.string :name
-      t.float :percentage
+      t.integer :percentage
       t.integer :quantity_threshold
 
       t.timestamps
